Refactored main window out of static context to separate class
